fixed errors

+++ crux_db/scripts/sub_blast1.sh
@@ -75,4 +75,10 @@ python ${DB}/scripts/combine_and_dereplicate_fasta.py -o ${ODIR}/${NAME}_BLAST/$
 
 printf "${FILE}_blast_1.complete" > ${QU}_blast_1.complete
 rm ${ODIR}/${NAME}_BLAST/${TYPE}_BLAST_out/raw/${FILE}_blast1_out.txt
-rm ${ODIR}/${NAME}_BLAST/${TYPE}_BLAST_out/fasta/${FILE}_blast1_out.fasta.tmp
+
+if [ -e "${ODIR}/${NAME}_BLAST/${TYPE}_BLAST_out/fasta/${FILE}_blast1_out.fasta.tmp" ];
+then
+  rm ${ODIR}/${NAME}_BLAST/${TYPE}_BLAST_out/fasta/${FILE}_blast1_out.fasta.tmp
+else
+  echo ""
+fi
